fix: fix failed test cases and move 'Entries' to 'kclvm-parser'.
