+++ kclvm/driver/src/test_data/test_vendor/.kpm/config/kpm.json
@@ -1,0 +1,1 @@
+{"DefaultOciRegistry":"ghcr.io","DefaultOciRepo":"kcl-lang","DefaultOciPlainHttp":false}
