+++ kclvm/parser/src/entry.rs
@@ -1,0 +1,413 @@
+use anyhow::Result;
+use kclvm_config::modfile::get_pkg_root;
+use kclvm_config::modfile::KCL_FILE_SUFFIX;
+use kclvm_config::path::ModRelativePath;
+use kclvm_utils::path::PathPrefix;
+use kclvm_utils::path::{is_absolute, is_dir, path_exist};
+use std::collections::VecDeque;
+
+use crate::LoadProgramOptions;
+
+/// [`Entries`] is a map of package name to package root path for one compilation
+/// # note
+///
+/// The [`entries`] in [`Entries`] is ordered, and the order of Entrys may affect the result of the compilation.
+/// The reason why the [`Entries`] is not an [`IndexMap`] is that the [`entries`] is duplicable and ordered.
+#[derive(Default, Debug)]
+pub struct Entries {
+    entries: VecDeque<Entry>,
+}
+
+impl Entries {
+    /// [`push_entry`] will push a new [`Entry`] into [`Entries`].
+    pub fn push_entry(&mut self, entry: Entry) {
+        self.entries.push_back(entry);
+    }
+
+    /// [`push`] will push a new [`Entry`] into [`Entries`] with the given name and path.
+    pub fn push(&mut self, name: String, path: String) {
+        self.entries.push_back(Entry::new(name, path));
+    }
+
+    /// [`contains_pkg_name`] will return [`Option::Some`] if there is an [`Entry`] with the given name in [`Entries`].
+    pub fn contains_pkg_name(&self, name: &String) -> Option<&Entry> {
+        self.entries.iter().find(|entry| entry.name() == name)
+    }
+
+    /// [`iter`] will return an iterator of [`Entry`] in [`Entries`].
+    pub fn iter(&self) -> std::collections::vec_deque::Iter<Entry> {
+        self.entries.iter()
+    }
+
+    /// [`len`] will return the length of [`Entries`].
+    pub fn len(&self) -> usize {
+        self.entries.len()
+    }
+
+    /// [`get_nth_entry`] will return the nth [`Entry`] in [`Entries`].
+    #[allow(dead_code)]
+    pub fn get_nth_entry(&self, n: usize) -> Option<&Entry> {
+        if n >= self.len() {
+            return None;
+        }
+        let mut count = 0;
+        for entry in self.entries.iter() {
+            if count == n {
+                return Some(entry);
+            }
+            count += 1;
+        }
+        return None;
+    }
+
+    /// [`get_nth_entry_by_name`] will return the nth [`Entry`] by name in [`Entries`].
+    pub fn get_nth_entry_by_name(&self, n: usize, name: &str) -> Option<&Entry> {
+        self.entries
+            .iter()
+            .enumerate()
+            .filter(|(_, entry)| name == entry.name())
+            .nth(n)
+            .map(|(_, entry)| entry)
+    }
+
+    /// [`apply_to_all_entries`] will apply the given function to all [`Entry`] in [`Entries`].
+    pub fn apply_to_all_entries<F>(&mut self, f: F) -> Result<(), String>
+    where
+        F: FnMut(&mut Entry) -> Result<(), String>,
+    {
+        self.entries.iter_mut().try_for_each(f)?;
+        return Ok(());
+    }
+}
+
+/// [`Entry`] is a package name and package root path pair.
+#[derive(Default, Debug)]
+pub struct Entry {
+    name: String,
+    path: String,
+    k_files: Vec<String>,
+}
+
+impl Entry {
+    /// [`new`] will create a new [`Entry`] with the given name and path.
+    pub fn new(name: String, path: String) -> Self {
+        Self {
+            name,
+            path,
+            k_files: vec![],
+        }
+    }
+
+    /// [`name`] will return the name of [`Entry`].
+    pub fn name(&self) -> &String {
+        &self.name
+    }
+
+    /// [`path`] will return the path of [`Entry`].
+    pub fn path(&self) -> &String {
+        &self.path
+    }
+
+    /// [`set_path`] will set the path of [`Entry`] to the given path.
+    pub fn set_path(&mut self, path: String) {
+        self.path = path;
+    }
+
+    /// [`extend_k_files`] will extend the k files of [`Entry`] to the given k file.
+    pub fn extend_k_files(&mut self, k_files: Vec<String>) {
+        self.k_files.extend(k_files);
+    }
+
+    /// [`get_k_files`] will return the k files of [`Entry`].
+    pub fn get_k_files(&self) -> &Vec<String> {
+        &self.k_files
+    }
+}
+
+/// [`get_compile_entries_from_paths`] returns all the [`Entries`] for compilation from the given [`file_paths`].
+///
+/// # Note
+/// If the path in [`file_paths`] is a normal path or a [`ModRelativePath`] with prefix `${KCL_MOD}`, the package will be named as `__main__`.
+/// If the path in [`file_paths`] is a [`ModRelativePath`], the package will be named by the suffix of [`ModRelativePath`].
+///
+/// # Error
+/// The package root path for package name `__main__` is only used once. If there are multiple
+/// package root paths for `__main__`, an error `conflict kcl.mod file` is returned.
+///
+/// # Example
+///
+/// ```rust
+/// use std::path::PathBuf;
+/// use kclvm_config::modfile::get_compile_entries_from_paths;
+/// let testpath = PathBuf::from("./src/testdata/multimods").canonicalize().unwrap();
+///
+/// // [`kcl1_path`] is a normal path of the package [`kcl1`] root directory.
+/// // It looks like `/xxx/xxx/xxx`.
+/// let kcl1_path = testpath.join("kcl1");
+///
+/// // [`kcl2_path`] is a mod relative path of the packege [`kcl2`] root directory.
+/// // It looks like `${kcl2:KCL_MOD}/xxx/xxx`
+/// let kcl2_path = PathBuf::from("${kcl2:KCL_MOD}/main.k");
+///
+/// // [`kcl3_path`] is a mod relative path of the [`__main__`] packege.
+/// let kcl3_path = PathBuf::from("${KCL_MOD}/main.k");
+///
+/// // [`external_pkgs`] is a map to show the real path of the mod relative path [`kcl2`].
+/// let mut external_pkgs = std::collections::HashMap::<String, String>::new();
+/// external_pkgs.insert("kcl2".to_string(), testpath.join("kcl2").to_str().unwrap().to_string());
+///
+/// // [`get_compile_entries_from_paths`] will return the map of package name to package root real path.
+/// let entries = get_compile_entries_from_paths(
+///     &[
+///         kcl1_path.to_str().unwrap().to_string(),
+///         kcl2_path.display().to_string(),
+///         kcl3_path.display().to_string(),
+///     ],
+/// external_pkgs).unwrap();
+///
+/// assert_eq!(entries.len(), 3);
+///
+/// assert_eq!(entries.get_nth_entry(0).unwrap().name(), "__main__");
+/// assert_eq!(
+///     PathBuf::from(entries.get_nth_entry(0).unwrap().path())
+///         .canonicalize()
+///         .unwrap()
+///         .display()
+///         .to_string(),
+///     kcl1_path.canonicalize().unwrap().to_str().unwrap()
+/// );
+///
+/// assert_eq!(entries.get_nth_entry(1).unwrap().name(), "kcl2");
+/// assert_eq!(
+///     PathBuf::from(entries.get_nth_entry(1).unwrap().path())
+///         .canonicalize()
+///         .unwrap()
+///         .display()
+///         .to_string(),
+///     testpath
+///         .join("kcl2")
+///         .canonicalize()
+///         .unwrap()
+///         .to_str()
+///         .unwrap()
+/// );
+///
+/// assert_eq!(entries.get_nth_entry(2).unwrap().name(), "__main__");
+/// assert_eq!(
+///     PathBuf::from(entries.get_nth_entry(2).unwrap().path())
+///         .display()
+///         .to_string(),
+///     kcl1_path.join("main.k").canonicalize().unwrap().to_str().unwrap()
+/// );
+///
+/// ```
+pub fn get_compile_entries_from_paths(
+    file_paths: &[String],
+    opts: &LoadProgramOptions,
+) -> Result<Entries, String> {
+    if file_paths.is_empty() {
+        return Err("No input KCL files or paths".to_string());
+    }
+    let mut result = Entries::default();
+    let mut m = std::collections::HashMap::<String, String>::new();
+    for s in file_paths {
+        let path = ModRelativePath::from(s.to_string());
+        // If the path is a [`ModRelativePath`] with preffix '${<package_name>:KCL_MOD}',
+        // calculate the real path and the package name.
+        if let Some((pkg_name, pkg_path)) = path
+            .get_root_pkg_name()
+            .map_err(|err| err.to_string())?
+            .and_then(|name| {
+                opts.package_maps
+                    .get(&name)
+                    .map(|pkg_path: &String| (name, pkg_path))
+            })
+        {
+            // Replace the mod relative path preffix '${<pkg_name>:KCL_MOD}' with the real path.
+            let s = path
+                .canonicalize_by_root_path(pkg_path)
+                .map_err(|err| err.to_string())?;
+            if let Some(root) = get_pkg_root(&s) {
+                let mut entry = Entry::new(pkg_name.clone(), root.clone());
+                entry.extend_k_files(get_main_files_from_pkg_path(&s, &root, &pkg_name, opts)?);
+                result.push_entry(entry);
+                continue;
+            }
+            // If the [`ModRelativePath`] with preffix '${KCL_MOD}'
+        } else if path.is_relative_path().map_err(|err| err.to_string())?
+            && path
+                .get_root_pkg_name()
+                .map_err(|err| err.to_string())?
+                .is_none()
+        {
+            // Push it into `result`, and deal it later.
+            result.push(kclvm_ast::MAIN_PKG.to_string(), path.get_path());
+            continue;
+        } else if let Some(root) = get_pkg_root(s) {
+            // If the path is a normal path.
+            m.insert(root.clone(), root.clone());
+            let mut entry = Entry::new(kclvm_ast::MAIN_PKG.to_string(), root.clone());
+            entry.extend_k_files(get_main_files_from_pkg_path(
+                &s,
+                &root,
+                &kclvm_ast::MAIN_PKG.to_string(),
+                opts,
+            )?);
+            result.push_entry(entry);
+        }
+    }
+
+    if m.is_empty() {
+        m.insert("".to_string(), "".to_string());
+        let mut entry = Entry::new(kclvm_ast::MAIN_PKG.to_string(), "".to_string());
+        for s in file_paths {
+            entry.extend_k_files(get_main_files_from_pkg_path(
+                s,
+                "",
+                &kclvm_ast::MAIN_PKG.to_string(),
+                opts,
+            )?);
+        }
+        result.push_entry(entry);
+    }
+
+    if m.len() == 1 {
+        // If the [`ModRelativePath`] with preffix '${KCL_MOD}'.
+        // Replace the '${KCL_MOD}' by the root path of package '__main__'.
+
+        // Calculate the root path of package '__main__'.
+        let pkg_root;
+        let main_entry = result
+            .get_nth_entry_by_name(0, kclvm_ast::MAIN_PKG)
+            .ok_or_else(|| format!("program entry not found in {:?}", file_paths))?;
+        pkg_root = main_entry.path().to_string();
+
+        // Replace the '${KCL_MOD}' of all the paths with package name '__main__'.
+        result.apply_to_all_entries(|entry| {
+            let path = ModRelativePath::from(entry.path().to_string());
+            if entry.name() == kclvm_ast::MAIN_PKG
+                && path.is_relative_path().map_err(|err| err.to_string())?
+            {
+                entry.set_path(pkg_root.to_string());
+                entry.extend_k_files(get_main_files_from_pkg_path(
+                    &path
+                        .canonicalize_by_root_path(&pkg_root)
+                        .map_err(|err| err.to_string())?,
+                    &pkg_root,
+                    &kclvm_ast::MAIN_PKG.to_string(),
+                    opts,
+                )?);
+            }
+            return Ok(());
+        })?;
+        return Ok(result);
+    }
+
+    // Filling the kcl program files (*.k) in each entry.
+
+    Err(format!("conflict kcl.mod file paths: {:?}", m))
+}
+
+/// Get files in the main package with the package root.
+fn get_main_files_from_pkg_path(
+    pkg_path: &str,
+    root: &str,
+    pkg_name: &str,
+    opts: &LoadProgramOptions,
+) -> Result<Vec<String>, String> {
+    // fix path
+    let mut path_list = Vec::new();
+    let mut s = pkg_path.to_string();
+
+    let path = ModRelativePath::from(s.to_string());
+
+    if path.is_relative_path().map_err(|e| e.to_string())? {
+        if let Some(name) = path.get_root_pkg_name().map_err(|e| e.to_string())? {
+            if name == pkg_name {
+                s = path
+                    .canonicalize_by_root_path(root)
+                    .map_err(|e| e.to_string())?;
+            }
+        } else if path.is_relative_path().map_err(|e| e.to_string())? {
+            return Err(format!("Can not find {} in the path: {}", s, root));
+        }
+    }
+    if !root.is_empty() && !is_absolute(s.as_str()) {
+        let p = std::path::Path::new(s.as_str());
+        if let Ok(x) = std::fs::canonicalize(p) {
+            s = x.adjust_canonicalization();
+        }
+    }
+
+    path_list.push(s);
+
+    // get k files
+    let mut k_files: Vec<String> = Vec::new();
+    for (i, path) in path_list.iter().enumerate() {
+        // read dir/*.k
+        if is_dir(path) {
+            if opts.k_code_list.len() > i {
+                return Err("Invalid code list".to_string());
+            }
+            //k_code_list
+            for s in get_dir_files(path)? {
+                k_files.push(s);
+            }
+            continue;
+        } else {
+            k_files.push(path.to_string());
+        }
+    }
+
+    if k_files.is_empty() {
+        return Err("No input KCL files".to_string());
+    }
+
+    // check all file exists
+    for (i, filename) in k_files.iter().enumerate() {
+        if i < opts.k_code_list.len() {
+            continue;
+        }
+
+        if !path_exist(filename.as_str()) {
+            return Err(format!(
+                "Cannot find the kcl file, please check the file path {}",
+                filename.as_str(),
+            ));
+        }
+    }
+    Ok(k_files)
+}
+
+/// Get file list in the directory.
+fn get_dir_files(dir: &str) -> Result<Vec<String>, String> {
+    if !std::path::Path::new(dir).exists() {
+        return Ok(Vec::new());
+    }
+
+    let mut list = Vec::new();
+
+    for path in std::fs::read_dir(dir).unwrap() {
+        let path = path.unwrap();
+        if !path
+            .file_name()
+            .to_str()
+            .unwrap()
+            .ends_with(KCL_FILE_SUFFIX)
+        {
+            continue;
+        }
+        if path.file_name().to_str().unwrap().ends_with("_test.k") {
+            continue;
+        }
+        if path.file_name().to_str().unwrap().starts_with('_') {
+            continue;
+        }
+
+        let s = format!("{}", path.path().display());
+        list.push(s);
+    }
+
+    list.sort();
+    Ok(list)
+}
